+++ tests/assembly.rs
@@ -107,7 +107,7 @@ fn instructions() {
             drv %a1 %b %x
             %v0 = var i8
             %v1 = load i8 %v0
-            store i8 42 %v0
+            store i8 %v0 42
             halt
         }
     "};